+++ src/master/idx_tag.h
@@ -1,0 +1,19 @@
+// Copyright (c) 2015, Baidu.com, Inc. All Rights Reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+#ifndef BAIDU_GALAXY_IDX_TAG_H
+#define BAIDU_GALAXY_IDX_TAG_H
+
+namespace baidu {
+namespace galaxy {
+
+struct addr_tag {};
+struct pod_id_tag {};
+struct id_tag{};
+struct name_tag{};
+struct target_tag{};
+
+}
+}
+#endif
+
