+++ src/proto/master.proto
@@ -72,6 +72,7 @@ message UserDescriptor {
 
 message SubmitJobRequest {
     optional JobDescriptor job = 1;
+    optional string sid = 2;
 }
 
 message SubmitJobResponse {
